KAFKA-16593: Extract common class.
